[portafly] extract accounts DataListTable
