+++ portafly/src/components/pages/accounts/index.ts
@@ -1,1 +1,2 @@
 export * from './listing'
+export * from './AccountsDataListTable'
